requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements/requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via starlette
click==8.1.7
    # via uvicorn
dnspython==2.6.1
    # via pymongo
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.0
    # via -r requirements/requirements.in
h11==0.14.0
    # via uvicorn
humps==0.2.2
    # via -r requirements/requirements.in
idna==3.10
    # via anyio
mysql-connector-python==9.0.0
    # via -r requirements/requirements.in
pydantic==2.9.2
    # via
    #   -r requirements/requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.5.2
    # via -r requirements/requirements.in
pymongo==4.9.1
    # via -r requirements/requirements.in
python-dotenv==1.0.1
    # via pydantic-settings
randomtimestamp==2.2
    # via -r requirements/requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.38.6
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   uvicorn
uvicorn==0.31.0
    # via -r requirements/requirements.in
